+++ zeebe/clients/java/pom.xml
@@ -269,7 +269,7 @@
               <!-- https://github.com/OpenAPITools/openapi-generator/blob/master/docs/generators/java.md -->
               <generatorName>java</generatorName>
               <inputSpec>${openapi.dir}/rest-api.yaml</inputSpec>
-              <modelPackage>io.camunda.zeebe.gateway.protocol.rest</modelPackage>
+              <modelPackage>io.camunda.zeebe.client.protocol.rest</modelPackage>
               <generateApis>false</generateApis>
               <generateApiTests>false</generateApiTests>
               <generateSupportingFiles>false</generateSupportingFiles>
