+++ zeebe/clients/java/src/main/java/io/camunda/zeebe/client/impl/command/AssignUserTaskCommandImpl.java
@@ -22,7 +22,7 @@ import io.camunda.zeebe.client.api.command.FinalCommandStep;
 import io.camunda.zeebe.client.api.response.AssignUserTaskResponse;
 import io.camunda.zeebe.client.impl.http.HttpClient;
 import io.camunda.zeebe.client.impl.http.HttpZeebeFuture;
-import io.camunda.zeebe.gateway.protocol.rest.UserTaskAssignmentRequest;
+import io.camunda.zeebe.client.protocol.rest.UserTaskAssignmentRequest;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 import org.apache.hc.client5.http.config.RequestConfig;
