+++ zeebe/clients/java/src/main/java/io/camunda/zeebe/client/impl/command/TopologyRequestImpl.java
@@ -106,7 +106,7 @@ public final class TopologyRequestImpl implements TopologyRequestStep1 {
         httpRequestConfig
             .setResponseTimeout(requestTimeout.toMillis(), TimeUnit.MILLISECONDS)
             .build(),
-        io.camunda.zeebe.gateway.protocol.rest.TopologyResponse.class,
+        io.camunda.zeebe.client.protocol.rest.TopologyResponse.class,
         TopologyImpl::new,
         result);
   }
