+++ zeebe/clients/java/src/main/java/io/camunda/zeebe/client/impl/command/UpdateUserTaskCommandImpl.java
@@ -22,8 +22,8 @@ import io.camunda.zeebe.client.api.command.UpdateUserTaskCommandStep1;
 import io.camunda.zeebe.client.api.response.UpdateUserTaskResponse;
 import io.camunda.zeebe.client.impl.http.HttpClient;
 import io.camunda.zeebe.client.impl.http.HttpZeebeFuture;
-import io.camunda.zeebe.gateway.protocol.rest.Changeset;
-import io.camunda.zeebe.gateway.protocol.rest.UserTaskUpdateRequest;
+import io.camunda.zeebe.client.protocol.rest.Changeset;
+import io.camunda.zeebe.client.protocol.rest.UserTaskUpdateRequest;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collections;
