+++ zeebe/clients/java/src/main/java/io/camunda/zeebe/client/impl/response/PartitionInfoImpl.java
@@ -18,10 +18,10 @@ package io.camunda.zeebe.client.impl.response;
 import io.camunda.zeebe.client.api.response.PartitionBrokerHealth;
 import io.camunda.zeebe.client.api.response.PartitionBrokerRole;
 import io.camunda.zeebe.client.api.response.PartitionInfo;
+import io.camunda.zeebe.client.protocol.rest.Partition.HealthEnum;
+import io.camunda.zeebe.client.protocol.rest.Partition.RoleEnum;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.Partition;
-import io.camunda.zeebe.gateway.protocol.rest.Partition.HealthEnum;
-import io.camunda.zeebe.gateway.protocol.rest.Partition.RoleEnum;
 import java.util.Arrays;
 
 public class PartitionInfoImpl implements PartitionInfo {
@@ -60,7 +60,7 @@ public class PartitionInfoImpl implements PartitionInfo {
     }
   }
 
-  public PartitionInfoImpl(final io.camunda.zeebe.gateway.protocol.rest.Partition httpPartition) {
+  public PartitionInfoImpl(final io.camunda.zeebe.client.protocol.rest.Partition httpPartition) {
 
     if (httpPartition.getPartitionId() == null) {
       throw new RuntimeException("Unexpected missing partition ID. A partition ID is required.");
